refactor to move TaskName details into task.go
